requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.5.0
celery[redis]==5.2.7
python-dotenv==1.0.0
drf-yasg==1.21.5
django-filter
django-celery-results==2.5.0